requirements.txt
Python 3.10.8

Package                      Version
---------------------------- ---------
cachetools                   5.3.1
certifi                      2023.7.22
charset-normalizer           3.2.0
cloudinary                   1.33.0
contourpy                    1.1.0
Django                       4.2.4
djangorestframework          3.14.0
flatbuffers                  23.5.26
gast                         0.4.0
google-auth                  2.22.0
google-auth-oauthlib         1.0.0
google-pasta                 0.2.0
grpcio                       1.56.2
h5py                         3.9.0
idna                         3.4
keras                        2.13.1
matplotlib                   3.7.2
numpy                        1.24.3
oauthlib                     3.2.2
opencv-python                4.8.0.74
opt-einsum                   3.3.0
packaging                    23.1
Pillow                       10.0.0
pip                          23.2.1
protobuf                     4.23.4
pyasn1                       0.5.0
pyasn1-modules               0.3.0
pyparsing                    3.0.9
python-dateutil              2.8.2
pytz                         2023.3
requests                     2.31.0
requests-oauthlib            1.3.1
rsa                          4.9
setuptools                   63.2.0
sqlparse                     0.4.4
tensorboard                  2.13.0
tensorboard-data-server      0.7.1
tensorflow                   2.13.0
tensorflow-estimator         2.13.0
tensorflow-intel             2.13.0
tensorflow-io-gcs-filesystem 0.31.0
tzdata                       2023.3
urllib3                      1.26.16
